fix:修复i18n失败问题
