+++ apiserver/httpserver/auth.go
@@ -23,6 +23,7 @@ import (
 	"github.com/emicklei/go-restful/v3"
 	"github.com/golang/protobuf/proto"
 
+	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	"github.com/polarismesh/polaris-server/common/utils"
 )
@@ -65,7 +66,7 @@ func (h *HTTPServer) GetAuthServer(ws *restful.WebService) error {
 
 // AuthStatus auth status
 func (h *HTTPServer) AuthStatus(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	checker := h.authServer.GetAuthChecker()
 
@@ -82,7 +83,7 @@ func (h *HTTPServer) AuthStatus(req *restful.Request, rsp *restful.Response) {
 
 // Login 登陆函数
 func (h *HTTPServer) Login(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	loginReq := &api.LoginRequest{}
 
@@ -97,7 +98,7 @@ func (h *HTTPServer) Login(req *restful.Request, rsp *restful.Response) {
 
 // CreateUsers 批量创建用户
 func (h *HTTPServer) CreateUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var users UserArr
 
@@ -116,7 +117,7 @@ func (h *HTTPServer) CreateUsers(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUser 更新用户
 func (h *HTTPServer) UpdateUser(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -131,7 +132,7 @@ func (h *HTTPServer) UpdateUser(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUserPassword 更新用户
 func (h *HTTPServer) UpdateUserPassword(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	user := &api.ModifyUserPassword{}
 
@@ -146,7 +147,7 @@ func (h *HTTPServer) UpdateUserPassword(req *restful.Request, rsp *restful.Respo
 
 // DeleteUsers 批量删除用户
 func (h *HTTPServer) DeleteUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var users UserArr
 
@@ -165,7 +166,7 @@ func (h *HTTPServer) DeleteUsers(req *restful.Request, rsp *restful.Response) {
 
 // GetUsers 查询用户
 func (h *HTTPServer) GetUsers(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -175,7 +176,7 @@ func (h *HTTPServer) GetUsers(req *restful.Request, rsp *restful.Response) {
 
 // GetUserToken 获取这个用户所关联的所有用户组列表信息，支持翻页
 func (h *HTTPServer) GetUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 	queryParams := utils.ParseQueryParams(req)
 
 	user := &api.User{
@@ -187,7 +188,7 @@ func (h *HTTPServer) GetUserToken(req *restful.Request, rsp *restful.Response) {
 
 // UpdateUserToken 更改用户的token
 func (h *HTTPServer) UpdateUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -202,7 +203,7 @@ func (h *HTTPServer) UpdateUserToken(req *restful.Request, rsp *restful.Response
 
 // ResetUserToken 重置用户 token
 func (h *HTTPServer) ResetUserToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	user := &api.User{}
 
@@ -217,7 +218,7 @@ func (h *HTTPServer) ResetUserToken(req *restful.Request, rsp *restful.Response)
 
 // CreateGroup 创建用户组
 func (h *HTTPServer) CreateGroup(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -232,7 +233,7 @@ func (h *HTTPServer) CreateGroup(req *restful.Request, rsp *restful.Response) {
 
 // UpdateGroups 更新用户组
 func (h *HTTPServer) UpdateGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var groups ModifyGroupArr
 
@@ -251,7 +252,7 @@ func (h *HTTPServer) UpdateGroups(req *restful.Request, rsp *restful.Response) {
 
 // DeleteGroups 删除用户组
 func (h *HTTPServer) DeleteGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var groups GroupArr
 
@@ -270,7 +271,7 @@ func (h *HTTPServer) DeleteGroups(req *restful.Request, rsp *restful.Response) {
 
 // GetGroups 获取用户组列表
 func (h *HTTPServer) GetGroups(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -280,7 +281,7 @@ func (h *HTTPServer) GetGroups(req *restful.Request, rsp *restful.Response) {
 
 // GetGroup 获取用户组详细
 func (h *HTTPServer) GetGroup(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -294,7 +295,7 @@ func (h *HTTPServer) GetGroup(req *restful.Request, rsp *restful.Response) {
 
 // GetGroupToken 获取用户组 token
 func (h *HTTPServer) GetGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -308,7 +309,7 @@ func (h *HTTPServer) GetGroupToken(req *restful.Request, rsp *restful.Response) 
 
 // UpdateGroupToken 更新用户组 token
 func (h *HTTPServer) UpdateGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -323,7 +324,7 @@ func (h *HTTPServer) UpdateGroupToken(req *restful.Request, rsp *restful.Respons
 
 // ResetGroupToken 重置用户组 token
 func (h *HTTPServer) ResetGroupToken(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	group := &api.UserGroup{}
 
@@ -338,7 +339,7 @@ func (h *HTTPServer) ResetGroupToken(req *restful.Request, rsp *restful.Response
 
 // CreateStrategy 创建鉴权策略
 func (h *HTTPServer) CreateStrategy(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	strategy := &api.AuthStrategy{}
 
@@ -353,7 +354,7 @@ func (h *HTTPServer) CreateStrategy(req *restful.Request, rsp *restful.Response)
 
 // UpdateStrategies 更新鉴权策略
 func (h *HTTPServer) UpdateStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var strategies ModifyStrategyArr
 
@@ -372,7 +373,7 @@ func (h *HTTPServer) UpdateStrategies(req *restful.Request, rsp *restful.Respons
 
 // DeleteStrategies 批量删除鉴权策略
 func (h *HTTPServer) DeleteStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var strategies StrategyArr
 
@@ -391,7 +392,7 @@ func (h *HTTPServer) DeleteStrategies(req *restful.Request, rsp *restful.Respons
 
 // GetStrategies 批量获取鉴权策略
 func (h *HTTPServer) GetStrategies(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -401,7 +402,7 @@ func (h *HTTPServer) GetStrategies(req *restful.Request, rsp *restful.Response) 
 
 // GetStrategy 获取鉴权策略详细
 func (h *HTTPServer) GetStrategy(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -415,7 +416,7 @@ func (h *HTTPServer) GetStrategy(req *restful.Request, rsp *restful.Response) {
 
 // GetPrincipalResources 获取鉴权策略详细
 func (h *HTTPServer) GetPrincipalResources(req *restful.Request, rsp *restful.Response) {
-	handler := &Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
