+++ apiserver/httpserver/server.go
@@ -31,6 +31,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/polarismesh/polaris-server/apiserver"
+	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
 	v1 "github.com/polarismesh/polaris-server/apiserver/httpserver/v1"
 	v2 "github.com/polarismesh/polaris-server/apiserver/httpserver/v2"
 	"github.com/polarismesh/polaris-server/auth"
@@ -523,7 +524,7 @@ func (h *HTTPServer) enterAuth(req *restful.Request, rsp *restful.Response) erro
 			zap.String("request-id", rid),
 			zap.String("platform-id", pid),
 			zap.String("platform-token", pToken))
-		HTTPResponse(req, rsp, api.NotAllowedAccess)
+		handler.HTTPResponse(req, rsp, api.NotAllowedAccess)
 		return errors.New("http access is not allowed")
 	}
 	return nil
@@ -547,7 +548,7 @@ func (h *HTTPServer) enterRateLimit(req *restful.Request, rsp *restful.Response)
 	if ok := h.rateLimit.Allow(plugin.IPRatelimit, segments[0]); !ok {
 		log.Error("ip ratelimit is not allow", zap.String("client", address),
 			zap.String("request-id", rid))
-		HTTPResponse(req, rsp, api.IPRateLimit)
+		handler.HTTPResponse(req, rsp, api.IPRateLimit)
 		return errors.New("ip ratelimit is not allow")
 	}
 
@@ -557,7 +558,7 @@ func (h *HTTPServer) enterRateLimit(req *restful.Request, rsp *restful.Response)
 	if ok := h.rateLimit.Allow(plugin.APIRatelimit, apiName); !ok {
 		log.Error("api ratelimit is not allow", zap.String("client", address),
 			zap.String("request-id", rid), zap.String("api", apiName))
-		HTTPResponse(req, rsp, api.APIRateLimit)
+		handler.HTTPResponse(req, rsp, api.APIRateLimit)
 		return errors.New("api ratelimit is not allow")
 	}
