+++ apiserver/httpserver/v1/naming_console_access.go
@@ -26,6 +26,7 @@ import (
 	"github.com/golang/protobuf/proto"
 	restfulspec "github.com/polarismesh/go-restful-openapi/v2"
 
+	"github.com/polarismesh/polaris-server/apiserver/httpserver/handler"
 	api "github.com/polarismesh/polaris-server/common/api/v1"
 	"github.com/polarismesh/polaris-server/common/utils"
 )
@@ -165,7 +166,7 @@ func (h *HTTPServerV1) addDefaultAccess(ws *restful.WebService) {
 
 // CreateNamespaces 创建命名空间
 func (h *HTTPServerV1) CreateNamespaces(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var namespaces NamespaceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -183,7 +184,7 @@ func (h *HTTPServerV1) CreateNamespaces(req *restful.Request, rsp *restful.Respo
 
 // DeleteNamespaces 删除命名空间
 func (h *HTTPServerV1) DeleteNamespaces(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var namespaces NamespaceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -207,7 +208,7 @@ func (h *HTTPServerV1) DeleteNamespaces(req *restful.Request, rsp *restful.Respo
 
 // UpdateNamespaces 修改命名空间
 func (h *HTTPServerV1) UpdateNamespaces(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var namespaces NamespaceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -231,7 +232,7 @@ func (h *HTTPServerV1) UpdateNamespaces(req *restful.Request, rsp *restful.Respo
 
 // GetNamespaces 查询命名空间
 func (h *HTTPServerV1) GetNamespaces(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	ret := h.namespaceServer.GetNamespaces(handler.ParseHeaderContext(), req.Request.URL.Query())
 	handler.WriteHeaderAndProto(ret)
@@ -239,7 +240,7 @@ func (h *HTTPServerV1) GetNamespaces(req *restful.Request, rsp *restful.Response
 
 // GetNamespaceToken 命名空间token的获取
 func (h *HTTPServerV1) GetNamespaceToken(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	token := req.HeaderParameter("Polaris-Token")
 	ctx := context.WithValue(context.Background(), utils.StringContext("polaris-token"), token)
@@ -256,7 +257,7 @@ func (h *HTTPServerV1) GetNamespaceToken(req *restful.Request, rsp *restful.Resp
 
 // UpdateNamespaceToken 更新命名空间的token
 func (h *HTTPServerV1) UpdateNamespaceToken(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var namespace api.Namespace
 	ctx, err := handler.Parse(&namespace)
@@ -271,7 +272,7 @@ func (h *HTTPServerV1) UpdateNamespaceToken(req *restful.Request, rsp *restful.R
 
 // CreateServices 创建服务
 func (h *HTTPServerV1) CreateServices(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var services ServiceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -289,7 +290,7 @@ func (h *HTTPServerV1) CreateServices(req *restful.Request, rsp *restful.Respons
 
 // DeleteServices 删除服务
 func (h *HTTPServerV1) DeleteServices(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var services ServiceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -313,7 +314,7 @@ func (h *HTTPServerV1) DeleteServices(req *restful.Request, rsp *restful.Respons
 
 // UpdateServices 修改服务
 func (h *HTTPServerV1) UpdateServices(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var services ServiceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -336,7 +337,7 @@ func (h *HTTPServerV1) UpdateServices(req *restful.Request, rsp *restful.Respons
 
 // GetServices 查询服务
 func (h *HTTPServerV1) GetServices(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ctx := handler.ParseHeaderContext()
@@ -346,14 +347,14 @@ func (h *HTTPServerV1) GetServices(req *restful.Request, rsp *restful.Response) 
 
 // GetServicesCount 查询服务总数
 func (h *HTTPServerV1) GetServicesCount(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 	ret := h.namingServer.GetServicesCount(handler.ParseHeaderContext())
 	handler.WriteHeaderAndProto(ret)
 }
 
 // GetServiceToken 获取服务token
 func (h *HTTPServerV1) GetServiceToken(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 	token := req.HeaderParameter("Polaris-Token")
 	ctx := context.WithValue(context.Background(), utils.StringContext("polaris-token"), token)
 
@@ -370,7 +371,7 @@ func (h *HTTPServerV1) GetServiceToken(req *restful.Request, rsp *restful.Respon
 
 // UpdateServiceToken 更新服务token
 func (h *HTTPServerV1) UpdateServiceToken(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var service api.Service
 	ctx, err := handler.Parse(&service)
@@ -384,7 +385,7 @@ func (h *HTTPServerV1) UpdateServiceToken(req *restful.Request, rsp *restful.Res
 
 // CreateServiceAlias service alias
 func (h *HTTPServerV1) CreateServiceAlias(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var alias api.ServiceAlias
 	ctx, err := handler.Parse(&alias)
@@ -398,7 +399,7 @@ func (h *HTTPServerV1) CreateServiceAlias(req *restful.Request, rsp *restful.Res
 
 // UpdateServiceAlias 修改服务别名
 func (h *HTTPServerV1) UpdateServiceAlias(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var alias api.ServiceAlias
 	ctx, err := handler.Parse(&alias)
@@ -418,7 +419,7 @@ func (h *HTTPServerV1) UpdateServiceAlias(req *restful.Request, rsp *restful.Res
 
 // DeleteServiceAliases 删除服务别名
 func (h *HTTPServerV1) DeleteServiceAliases(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var aliases ServiceAliasArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -441,7 +442,7 @@ func (h *HTTPServerV1) DeleteServiceAliases(req *restful.Request, rsp *restful.R
 
 // GetServiceAliases 根据源服务获取服务别名
 func (h *HTTPServerV1) GetServiceAliases(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetServiceAliases(handler.ParseHeaderContext(), queryParams)
@@ -450,7 +451,7 @@ func (h *HTTPServerV1) GetServiceAliases(req *restful.Request, rsp *restful.Resp
 
 // CreateInstances 创建服务实例
 func (h *HTTPServerV1) CreateInstances(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var instances InstanceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -468,7 +469,7 @@ func (h *HTTPServerV1) CreateInstances(req *restful.Request, rsp *restful.Respon
 
 // DeleteInstances 删除服务实例
 func (h *HTTPServerV1) DeleteInstances(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var instances InstanceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -492,7 +493,7 @@ func (h *HTTPServerV1) DeleteInstances(req *restful.Request, rsp *restful.Respon
 
 // DeleteInstancesByHost 根据host删除服务实例
 func (h *HTTPServerV1) DeleteInstancesByHost(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var instances InstanceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -516,7 +517,7 @@ func (h *HTTPServerV1) DeleteInstancesByHost(req *restful.Request, rsp *restful.
 
 // UpdateInstances 修改服务实例
 func (h *HTTPServerV1) UpdateInstances(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var instances InstanceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -540,7 +541,7 @@ func (h *HTTPServerV1) UpdateInstances(req *restful.Request, rsp *restful.Respon
 
 // UpdateInstancesIsolate 修改服务实例的隔离状态
 func (h *HTTPServerV1) UpdateInstancesIsolate(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var instances InstanceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -564,7 +565,7 @@ func (h *HTTPServerV1) UpdateInstancesIsolate(req *restful.Request, rsp *restful
 
 // GetInstances 查询服务实例
 func (h *HTTPServerV1) GetInstances(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetInstances(handler.ParseHeaderContext(), queryParams)
@@ -573,7 +574,7 @@ func (h *HTTPServerV1) GetInstances(req *restful.Request, rsp *restful.Response)
 
 // GetInstancesCount 查询服务实例
 func (h *HTTPServerV1) GetInstancesCount(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	ret := h.namingServer.GetInstancesCount(handler.ParseHeaderContext())
 	handler.WriteHeaderAndProto(ret)
@@ -581,7 +582,7 @@ func (h *HTTPServerV1) GetInstancesCount(req *restful.Request, rsp *restful.Resp
 
 // CreateRoutings 创建规则路由
 func (h *HTTPServerV1) CreateRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -600,7 +601,7 @@ func (h *HTTPServerV1) CreateRoutings(req *restful.Request, rsp *restful.Respons
 
 // DeleteRoutings 删除规则路由
 func (h *HTTPServerV1) DeleteRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -624,7 +625,7 @@ func (h *HTTPServerV1) DeleteRoutings(req *restful.Request, rsp *restful.Respons
 
 // UpdateRoutings 修改规则路由
 func (h *HTTPServerV1) UpdateRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var routings RoutingArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -648,7 +649,7 @@ func (h *HTTPServerV1) UpdateRoutings(req *restful.Request, rsp *restful.Respons
 
 // GetRoutings 查询规则路由
 func (h *HTTPServerV1) GetRoutings(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetRoutingConfigs(handler.ParseHeaderContext(), queryParams)
@@ -657,7 +658,7 @@ func (h *HTTPServerV1) GetRoutings(req *restful.Request, rsp *restful.Response) 
 
 // CreateRateLimits 创建限流规则
 func (h *HTTPServerV1) CreateRateLimits(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var rateLimits RateLimitArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -675,7 +676,7 @@ func (h *HTTPServerV1) CreateRateLimits(req *restful.Request, rsp *restful.Respo
 
 // DeleteRateLimits 删除限流规则
 func (h *HTTPServerV1) DeleteRateLimits(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var rateLimits RateLimitArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -698,7 +699,7 @@ func (h *HTTPServerV1) DeleteRateLimits(req *restful.Request, rsp *restful.Respo
 
 // EnableRateLimits 激活限流规则
 func (h *HTTPServerV1) EnableRateLimits(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 	var rateLimits RateLimitArr
 	ctx, err := handler.ParseArray(func() proto.Message {
 		msg := &api.Rule{}
@@ -720,7 +721,7 @@ func (h *HTTPServerV1) EnableRateLimits(req *restful.Request, rsp *restful.Respo
 
 // UpdateRateLimits 修改限流规则
 func (h *HTTPServerV1) UpdateRateLimits(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var rateLimits RateLimitArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -744,7 +745,7 @@ func (h *HTTPServerV1) UpdateRateLimits(req *restful.Request, rsp *restful.Respo
 
 // GetRateLimits 查询限流规则
 func (h *HTTPServerV1) GetRateLimits(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetRateLimits(handler.ParseHeaderContext(), queryParams)
@@ -753,7 +754,7 @@ func (h *HTTPServerV1) GetRateLimits(req *restful.Request, rsp *restful.Response
 
 // CreateCircuitBreakers 创建熔断规则
 func (h *HTTPServerV1) CreateCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var circuitBreakers CircuitBreakerArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -772,7 +773,7 @@ func (h *HTTPServerV1) CreateCircuitBreakers(req *restful.Request, rsp *restful.
 
 // CreateCircuitBreakerVersions 创建熔断规则版本
 func (h *HTTPServerV1) CreateCircuitBreakerVersions(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var circuitBreakers CircuitBreakerArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -790,7 +791,7 @@ func (h *HTTPServerV1) CreateCircuitBreakerVersions(req *restful.Request, rsp *r
 
 // DeleteCircuitBreakers 删除熔断规则
 func (h *HTTPServerV1) DeleteCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var circuitBreakers CircuitBreakerArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -813,7 +814,7 @@ func (h *HTTPServerV1) DeleteCircuitBreakers(req *restful.Request, rsp *restful.
 
 // UpdateCircuitBreakers 修改熔断规则
 func (h *HTTPServerV1) UpdateCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var circuitBreakers CircuitBreakerArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -836,7 +837,7 @@ func (h *HTTPServerV1) UpdateCircuitBreakers(req *restful.Request, rsp *restful.
 
 // ReleaseCircuitBreakers 发布熔断规则
 func (h *HTTPServerV1) ReleaseCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var configRelease ConfigReleaseArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -859,7 +860,7 @@ func (h *HTTPServerV1) ReleaseCircuitBreakers(req *restful.Request, rsp *restful
 
 // UnBindCircuitBreakers 解绑熔断规则
 func (h *HTTPServerV1) UnBindCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var configRelease ConfigReleaseArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -882,7 +883,7 @@ func (h *HTTPServerV1) UnBindCircuitBreakers(req *restful.Request, rsp *restful.
 
 // GetCircuitBreaker 根据id和version获取熔断规则
 func (h *HTTPServerV1) GetCircuitBreaker(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetCircuitBreaker(handler.ParseHeaderContext(), queryParams)
@@ -891,7 +892,7 @@ func (h *HTTPServerV1) GetCircuitBreaker(req *restful.Request, rsp *restful.Resp
 
 // GetCircuitBreakerVersions 查询熔断规则的所有版本
 func (h *HTTPServerV1) GetCircuitBreakerVersions(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetCircuitBreakerVersions(handler.ParseHeaderContext(), queryParams)
@@ -900,7 +901,7 @@ func (h *HTTPServerV1) GetCircuitBreakerVersions(req *restful.Request, rsp *rest
 
 // GetMasterCircuitBreakers 查询master熔断规则
 func (h *HTTPServerV1) GetMasterCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetMasterCircuitBreakers(handler.ParseHeaderContext(), queryParams)
@@ -909,7 +910,7 @@ func (h *HTTPServerV1) GetMasterCircuitBreakers(req *restful.Request, rsp *restf
 
 // GetReleaseCircuitBreakers 根据规则id查询已发布的熔断规则
 func (h *HTTPServerV1) GetReleaseCircuitBreakers(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetReleaseCircuitBreakers(handler.ParseHeaderContext(), queryParams)
@@ -918,7 +919,7 @@ func (h *HTTPServerV1) GetReleaseCircuitBreakers(req *restful.Request, rsp *rest
 
 // GetCircuitBreakerByService 根据服务查询绑定熔断规则
 func (h *HTTPServerV1) GetCircuitBreakerByService(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	queryParams := utils.ParseQueryParams(req)
 	ret := h.namingServer.GetCircuitBreakerByService(handler.ParseHeaderContext(), queryParams)
@@ -927,7 +928,7 @@ func (h *HTTPServerV1) GetCircuitBreakerByService(req *restful.Request, rsp *res
 
 // GetServiceOwner 根据服务获取服务负责人
 func (h *HTTPServerV1) GetServiceOwner(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 
 	var services ServiceArr
 	ctx, err := handler.ParseArray(func() proto.Message {
@@ -945,7 +946,7 @@ func (h *HTTPServerV1) GetServiceOwner(req *restful.Request, rsp *restful.Respon
 
 // GetCircuitBreakerToken 获取熔断规则token
 func (h *HTTPServerV1) GetCircuitBreakerToken(req *restful.Request, rsp *restful.Response) {
-	handler := &utils.Handler{req, rsp}
+	handler := &handler.Handler{req, rsp}
 	token := req.HeaderParameter("Polaris-Token")
 	ctx := context.WithValue(context.Background(), utils.StringContext("polaris-token"), token)
